backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-socketio==5.3.6
flask-cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyjwt==2.9.0
redis==5.0.8
requests==2.32.3
gunicorn==22.0.0
eventlet==0.36.1
pytest==8.3.5
